+++ lib/eval_annotation.dart
@@ -3,3 +3,4 @@ library eval_annotation;
 
 export 'src/runtime_override.dart';
 export 'src/bind.dart';
+export 'src/permission.dart';