+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.7.1
+- Added `@UnionOf` and `@AssertPermission` annotations.
+- Added `overrideLibrary` parameter to `@Bind` annotation.
+
 ## 0.7.0
 - Added `@Bind` annotation for the dart_eval binding generator.
 - Support for Dart 3